+++ gatsby-node.js
@@ -33,11 +33,12 @@ exports.createPages = ({ actions, graphql }) => {
 
     posts.forEach(edge => {
       const id = edge.node.id
+      console.log(36, edge.node.frontmatter.templateKey)
       createPage({
         path: edge.node.fields.slug,
         tags: edge.node.frontmatter.tags,
         component: path.resolve(
-          `src/templates/${String(edge.node.frontmatter.templateKey)}.tsx`
+          `src/containers/${String(edge.node.frontmatter.templateKey)}Container.tsx`
         ),
         // additional data can be passed via context
         context: {
@@ -63,7 +64,7 @@ exports.createPages = ({ actions, graphql }) => {
 
       createPage({
         path: tagPath,
-        component: path.resolve(`src/templates/tags.js`),
+        component: path.resolve(`src/containers/tags.js`),
         context: {
           tag,
         },
