+++ src/cms/preview-templates/BlogPostPreview.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { BlogPostTemplate } from '../../templates/blog-post'
+import { BlogPostTemplate } from '../../containers/blogContainer'
 
 const BlogPostPreview = ({ entry, widgetFor }) => (
   <BlogPostTemplate
