+++ src/pages/blog/2017-01-04-a-beginners-guide-to-brewing-with-chemex.md
@@ -1,5 +1,5 @@
 ---
-templateKey: blog-post
+templateKey: blog
 title: A beginners’ guide to brewing with Chemex
 date: 2017-01-04T15:04:10.000Z
 featuredpost: false
