+++ src/pages/blog/2017-01-04-just-in-small-batch-of-jamaican-blue-mountain-in-store-next-week.md
@@ -1,5 +1,5 @@
 ---
-templateKey: 'blog-post'
+templateKey: blog
 title: 'Just in: small batch of Jamaican Blue Mountain in store next week'
 date: 2017-01-04T15:04:10.000Z
 featuredpost: true
